pyproject.toml
[project]
name = "weebiee-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "numpy>=2.3.0",
    "tqdm>=4.67.1",
]
